Create README.md
This is an adapter to use N-Connectors on the LiteVNA. I did not want to use any screws to hold the LiteVNA in place. It is made such that the VNA fits tight in the adapter and by just unscrewing the two SMA connectors and sliding the VNA away from the connectors it can be lifted up from the frame and removed easily.
The reason for this solution is that I mostly use the VNA directly to the SMA connectors.